+++ esp-hal/src/rsa/esp32cX.rs
@@ -1,4 +1,4 @@
-use core::{convert::Infallible, marker::PhantomData, ptr::copy_nonoverlapping};
+use core::{convert::Infallible, marker::PhantomData};
 
 use crate::rsa::{
     implement_op,
@@ -32,7 +32,7 @@ impl<'d, DM: crate::Mode> Rsa<'d, DM> {
         }
     }
 
-    fn write_mode(&mut self, mode: u32) {
+    pub(super) fn write_mode(&mut self, mode: u32) {
         self.rsa.mode().write(|w| unsafe { w.bits(mode) });
     }
 
@@ -122,10 +122,6 @@ impl<'d, DM: crate::Mode> Rsa<'d, DM> {
     /// Checks if the RSA peripheral is idle.
     pub(super) fn is_idle(&mut self) -> bool {
         self.rsa.query_idle().read().query_idle().bit_is_set()
-    }
-
-    unsafe fn write_multi_operand_b<const N: usize>(&mut self, operand_b: &[u32; N]) {
-        copy_nonoverlapping(operand_b.as_ptr(), self.rsa.z_mem(0).as_ptr().add(N), N);
     }
 }
 
@@ -240,33 +236,6 @@ impl<'a, 'd, T: RsaMode, DM: crate::Mode, const N: usize> RsaModularExponentiati
 where
     T: RsaMode<InputType = [u32; N]>,
 {
-    /// Creates an instance of `RsaModularExponentiation`.
-    ///
-    /// `m_prime` could be calculated using `-(modular multiplicative inverse of
-    /// modulus) mod 2^32`.
-    ///
-    /// For more information refer to 19.3.1 of <https://www.espressif.com/sites/default/files/documentation/esp32-c3_technical_reference_manual_en.pdf>.
-    pub fn new(
-        rsa: &'a mut Rsa<'d, DM>,
-        exponent: &T::InputType,
-        modulus: &T::InputType,
-        m_prime: u32,
-    ) -> Self {
-        Self::set_mode(rsa);
-        unsafe {
-            rsa.write_operand_b(exponent);
-            rsa.write_modulus(modulus);
-        }
-        rsa.write_mprime(m_prime);
-        if rsa.is_search_enabled() {
-            rsa.write_search_position(Self::find_search_pos(exponent));
-        }
-        Self {
-            rsa,
-            phantom: PhantomData,
-        }
-    }
-
     fn find_search_pos(exponent: &T::InputType) -> u32 {
         for (i, byte) in exponent.iter().rev().enumerate() {
             if *byte == 0 {
@@ -278,13 +247,8 @@ where
     }
 
     /// Sets the modular exponentiation mode for the RSA hardware.
-    pub(super) fn set_mode(rsa: &mut Rsa<'d, DM>) {
+    fn write_mode(rsa: &mut Rsa<'d, DM>) {
         rsa.write_mode((N - 1) as u32)
-    }
-
-    /// Starts the modular exponentiation operation on the RSA hardware.
-    pub(super) fn start(&mut self) {
-        self.rsa.write_modexp_start();
     }
 }
 
@@ -298,39 +262,39 @@ where
 
     /// Creates an instance of `RsaModularMultiplication`.
     ///
-    /// `m_prime` can be calculated using `-(modular multiplicative inverse of
-    /// modulus) mod 2^32`.
+    /// - `r` can be calculated using `2 ^ ( bitlength * 2 ) mod modulus`.
+    /// - `m_prime` can be calculated using `-(modular multiplicative inverse of
+    ///   modulus) mod 2^32`.
     ///
     /// For more information refer to 19.3.1 of <https://www.espressif.com/sites/default/files/documentation/esp32-c3_technical_reference_manual_en.pdf>.
     pub fn new(
         rsa: &'a mut Rsa<'d, DM>,
         operand_a: &T::InputType,
-        operand_b: &T::InputType,
         modulus: &T::InputType,
+        r: &T::InputType,
         m_prime: u32,
     ) -> Self {
         Self::write_mode(rsa);
         rsa.write_mprime(m_prime);
-        unsafe {
-            rsa.write_modulus(modulus);
-            rsa.write_operand_a(operand_a);
-            rsa.write_operand_b(operand_b);
-        }
+        rsa.write_modulus(modulus);
+        rsa.write_operand_a(operand_a);
+        rsa.write_r(r);
+
         Self {
             rsa,
             phantom: PhantomData,
         }
     }
 
+    fn write_mode(rsa: &mut Rsa<'d, DM>) {
+        rsa.write_mode((N - 1) as u32)
+    }
+
     /// Starts the modular multiplication operation.
     ///
-    /// `r` could be calculated using `2 ^ ( bitlength * 2 ) mod modulus`.
-    ///
     /// For more information refer to 19.3.1 of <https://www.espressif.com/sites/default/files/documentation/esp32-c3_technical_reference_manual_en.pdf>.
-    pub fn start_modular_multiplication(&mut self, r: &T::InputType) {
-        unsafe {
-            self.rsa.write_r(r);
-        }
+    pub fn start_modular_multiplication(&mut self, operand_b: &T::InputType) {
+        self.rsa.write_operand_b(operand_b);
         self.start();
     }
 
@@ -343,28 +307,14 @@ impl<'a, 'd, T: RsaMode + Multi, DM: crate::Mode, const N: usize> RsaMultiplicat
 where
     T: RsaMode<InputType = [u32; N]>,
 {
-    /// Creates an instance of `RsaMultiplication`.
-    pub fn new(rsa: &'a mut Rsa<'d, DM>, operand_a: &T::InputType) -> Self {
-        Self::set_mode(rsa);
-        unsafe {
-            rsa.write_operand_a(operand_a);
-        }
-        Self {
-            rsa,
-            phantom: PhantomData,
-        }
-    }
-
     /// Starts the multiplication operation.
     pub fn start_multiplication(&mut self, operand_b: &T::InputType) {
-        unsafe {
-            self.rsa.write_multi_operand_b(operand_b);
-        }
+        self.rsa.write_multi_operand_b(operand_b);
         self.start();
     }
 
     /// Sets the multiplication mode for the RSA hardware.
-    pub(super) fn set_mode(rsa: &mut Rsa<'d, DM>) {
+    fn write_mode(rsa: &mut Rsa<'d, DM>) {
         rsa.write_mode((N * 2 - 1) as u32)
     }